Organize abstract classes
