+++ causal_networkx/metrics.py
@@ -1,3 +1,7 @@
+import numpy as np
+from sklearn.metrics import confusion_matrix
+
+
 def compare_networks(
     true_graph,
     pred_graph,
@@ -17,23 +21,50 @@ def compare_networks(
     cm : np.ndarray of shape (2, 2)
         The confusion matrix.
     """
-    # get all true edges
-    # y_true = true_graph.edges
-    # y_pred = pred_graph.edges
+    assert list(true_graph.nodes) == list(pred_graph.nodes)
 
-    # compute a score
-    # scores =
-    pass
+    # convert graphs to adjacency matrix in numpy array format
+    true_adj_mat = true_graph.to_adjacency_graph()
+    pred_adj_mat = pred_graph.to_adjacency_graph()
+
+    true_adj_mat = true_adj_mat > 0
+    pred_adj_mat = pred_adj_mat > 0
+
+    # vectorize
+    y_true = true_adj_mat.flatten()
+    y_pred = pred_adj_mat.flatten()
+
+    # compute the confusion matrix
+    conf_mat = confusion_matrix(y_true, y_pred)
+    return conf_mat
 
 
-def structure_hamming_dist(graph, other_graph):
+def structure_hamming_dist(graph, other_graph, double_for_anticausal: bool = True):
     """Compute structural hamming distance.
 
     Parameters
     ----------
     graph : _type_
-        _description_
+        Reference graph.
     other_graph : _type_
-        _description_
+        Other graph.
+    double_for_anticausal : bool, optional
+        Whether to count incorrect orientations as two mistakes, by default True
+
+    Returns
+    -------
+    shm : float
+        The hamming distance between 0 and infinity.
     """
-    pass
+    # convert graphs to adjacency matrix in numpy array format
+    adj_mat = graph.to_adjacency_graph()
+    other_adj_mat = other_graph.to_adjacency_graph()
+
+    diff = np.abs(adj_mat - other_adj_mat)
+
+    if double_for_anticausal:
+        return np.sum(diff)
+    else:
+        diff = diff + diff.transpose()
+        diff[diff > 1] = 1  # Ignoring the double edges.
+        return np.sum(diff) / 2
