+++ causal_networkx/scm.py
@@ -4,7 +4,7 @@ from typing import Callable, Dict
 
 import pandas as pd
 
-from causal_networkx.graphs.cgm import ADMG
+from causal_networkx import ADMG
 
 
 class StructuralCausalModel:
