+++ causal_networkx/tsgraph.py
@@ -1,4 +1,4 @@
-from causal_networkx.graphs.cgm import ADMG
+from causal_networkx import ADMG
 
 
 class TimeSeriesADMG(ADMG):
